Rename with .hpp & only put funct decl
